+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### :racing_car: Performance
 
 - ~20% faster VCR cassette recording through output sanitization caching.
+- Cache cURL commands.
 
 ## [4.4.4](https://github.com/schemathesis/schemathesis/compare/v4.4.3...v4.4.4) - 2025-11-08